Extract 'API' to a dedicated module
